Use sources as type deifnitions
Right now our build and CI take a lot of time. 1:40min - 2min on CI and
1min locally.

At the moment in packages anything outside of src folder is not checked
because of rootDir.

We have to generate "references" field on precommit.

All these issues can be avoided if remove .d.ts generation completely
and specify sources in "types" field. This way we get zero build time
for types and can switch to other build tool (like esbuild) for .js
generation and get almost instant build on vercel, on main CI and on publish.

Downsides

- it can be slower (actually not really caz we delete .tsbuildinfo cache
  anyway and we can still generate it with incremental flag)
- it contraditcs recommended practices for published packages (we can
  generate .d.ts in publish script only and use "types" as source
  internally)