Lift props state to designer
This is the prototype for the idea to manage all actions on designer
side and send to canvas same patches we send to server.

TODO
- remove props from insertInstance

Patches are handled naively and published from server queue so not
applied instantly. Though with keeping actual props store on designer
side we can remove local state and use atom.