Fix hover outline (#689)
Found an issue when hover a few elements fast and then move cursor out
of iframe hover gets stuck on some element. This is because we use
combination of mouseover and mouseout events.

So when we move out of one instance to another it looks like this

over
out
over
out

Which may lead to race conditions. So here out is canceled by every
hover and delayed a little to always emit last.