Update store in style panel directly
Before we sent messages from designer to canvas, patched data
and sent whole data back to designner.

This created latency so we had to keep local state and synchronize it
with effects.

Here I patched store and used its state in designer which is same as
local state. So we no longer need effect synchronization (and bugs
because of it).