Use instances index to compute selected instance (#839)
Because we tracked the tree and selected instance, traversed and updated
the value on every change which usually lead to running same code twice
in a row.

Here I added computed index for instances which can be reused in other
parts. It is generated by traversing tree. Later will be replaced with
for loop.

This allows us to store only selected id and compute instance with
`instancesById.get(selectedInstanceId)`.

This store need to be synchronized with designer to use it instead of
selected instance data in almost all cases.