Track delete shortcut on both canvas and designer (#947)
Here's the issue. We track shortcuts on canvas and designer though
designer invoke hotkeys with shared configuration and loose edge cases.
For example delete should not be invoked on inputs though shared
configuration enforce it. If we disable inputs then undo/redo will stop
working.

So for better safety we should track shortcuts in both designer and
canvas with the same code and update synchronized stores instead of
synchronizing shortcuts.