Compute indicator placement separately from drop target
We need to render placement indicator on both tree and canvas at the
same time. To achieve this we need to be able to compute position from
drop target which should be universal for tree and canvas.

Here introduced computeaIndicatorPlacement utility which accepts info
inferred from cursor but universal for realms.