Upload file instead of form data
Form data is complex format to work with and web api does not support
its stream parsing and we had to use remix utilitties.

Here upload file as is so node or worker gets stream and do anything
with it. So this is one less dependency  on node.

These streams are still bufferred but we can improve this at least for
images later to not exceed vercel memory.