Check types outside of src folder
We generate dts type for each package on publish.
It requires us to include only "src" to output same structure
as in src folder. Though we have configs and cli scripts outside of it
which should be typechecked, they cannot be ignored.

Ideally would be to have one-line commands in package.json.
Though typescript always has problems with build step.
In this case the CLI have everything to configure this except
"include" and "exclude" flags.

So here to workaround that I added tsconfig.dts.json to distinct dts
generation and typechecking.

Blocker for https://github.com/webstudio-is/webstudio-builder/pull/1634