Add expression utilities
Ref https://github.com/webstudio-is/webstudio-builder/issues/1529

Here added two utilities:

validateExpression to parse syntax and replace identifier.
For parsing used jsep package which is very limited subset of js.

executeExpression to execute the chain of expressions.
The code is unoptimized and will not cache results between calls.
But still should be fast enough because expressions are indented
to be quite simple.