Support copy pasting of data sources
Here fixed the case when copy pasted form looses state
and give errors on submit.

Immerhin now converted to esm to work in tests.
Added tests for instances copy paste plugin.
The new logic is copy data sources as is if they are bound
to copied instances.

The new computeExpressionsDependencies used to detect if expression
depends on data sources outside of copied instances scope.