Migrate to custom jest transform
Esbuild is not officially supported by jest
and all available packages do not fit all our need.

esbuild-jest we used for a long time have problems too.

The most important one is frigile heuristik to instrument
any code with `ock(` in text. And we have some `pointerLock(`
in design-system.

The other problem is instrumentation converts to commonjs
and result goes to esbuild which tries to convert back to esm
but with only default export. So jest fails anywhere
we try to import design-system in tests.

The other problem is lack of maintanance and very old packages inside.

Here refactored original package and removed everything we don't need
including mocks instrumentation.

This should fix the build in https://github.com/webstudio-is/webstudio-builder/pull/1986