Drag on canvas closest detachable
Before when start dragging we searched for closest rich editor
under cursor and did nothing when the instance is not detachable
(for example select value or switch indicator).

Here I changed the behavior to also look for closest detachable
instance and drag whole thing. This way we do not need toast like
in navigator and can canvas dnd just work.

Also fixed placeholder in "Select Value" component. Before was no way to
change it.