feat: add basic code editor
Here added code editor with variables completion support
for bindings UI. Simple textarea would not work because
same names can be used for different variables
to support seamless dnd for user. `$` is used to trigger
completion, in the future can be replaced with tab or
something like this. Current implementation is mostly
a copy paste from lexical playground.