feat: reimplement expression editor with codemirror
Ref https://github.com/webstudio-is/webstudio/issues/2532

Here introduce the new expression editor as a replacement
to code editor. It is implemented with codemirror and will
improve DX of our bindings and cms usage.

- autocomplete now completes nested objects in variables
- javascript syntax is highlighted in expressions
- json values are formatted as multiline text
- autocomplete shows fields value/type

Additionally improved a couple of things

- json objects are wrapped with parentheses `({ param: "value" })`
  to be treated as object instead of block in highlighter
- objects and arrays are rendered as "object" and "array"
  in variables list for cleaner UI

Integrated in builder only for variable value and json prop value.
Will require additional logic to integrate as expression editor.