refactor: add nesting rule to css engine
Nesting rule is universal representation for
instance style with support for nested state selectors
and nested media queries. Css is generated css of course
without css nesting and optimize media queries but in the future
may be generated with nesting for simplicity.

This is essential to support mixins in css engine.
Will support it in atomic css in other PRs.