chore: Protect session cookies (#4034)
## Description

Cookie Overwriting in Canvas:
Scripts running inside the canvas can add any cookie client-side. There
is no way on the server to distinguish whether a cookie was created on
the server or client-side. A bad actor can even write a signed cookie by
copying it from their own session, making it difficult to identify the
source of the cookie.

To protect against cookie overwrites, we set a null session cookie if a
user is using an authToken.
This ensures that any existing HttpOnly, secure session cookies cannot
be overwritten by client-side scripts
See Storage model
https://datatracker.ietf.org/doc/html/rfc6265#section-5.3
If the cookie store contains a cookie with the same name,
domain, and path as the newly created cookie:
...
If the newly created cookie was received from a "non-HTTP"
 API and the old-cookie's http-only-flag is set, abort these
 steps and ignore the newly created cookie entirely.

_same domain and path are enforced by `__Host-` prefix__

## Steps for reproduction

1. click button
2. expect xyz

## Code Review

- [ ] hi @kof, I need you to do
  - conceptual review (architecture, feature-correctness)
  - detailed review (read every line)
  - test it on preview

## Before requesting a review

- [ ] made a self-review
- [ ] added inline comments where things may be not obvious (the "why",
not "what")

## Before merging

- [ ] tested locally and on preview environment (preview dev login:
5de6)
- [ ] updated [test
cases](https://github.com/webstudio-is/webstudio/blob/main/apps/builder/docs/test-cases.md)
document
- [ ] added tests
- [ ] if any new env variables are added, added them to `.env` file