build: automatically publish CLI along with release (#4523)
We often forget to publish CLI packages after release and get some bug
reports from users because of it.

Here invoked publishing along with release generation when tag is
pushed.

The version is taken from release name and
will no longer be committed to repo.

Gonna test it once merged.