refactor: add templates meta support
Here added one more description file for components `*.template.tsx`.

It is decoupled from component meta to fix the issue with some
templates like Sheet being treated as components though we don't have
actual components with this name.

Another reason is have it separately to not include builder specific
code and dependencies into CLI and published websites. Template package
is defined as optional in peer dependencies meta.

In this PR migrated only Button and List as examples.
There is still not support for data sources and styles.