feat: allow saving expressions with unset variables
Ref https://github.com/webstudio-is/webstudio/issues/4768 https://github.com/webstudio-is/webstudio/issues/4753

User can paste expression with not defined variable
and should get it saved in case of accidental closing popover.

Now such expressions can be executed in builder and will be compiled
with undefined instead of variables in published sites.

In the future we need to add global diagnostics to let users find such
places faster.

Also fixed a couple of ui issues
- react error about using button inside of button in data variables list
- react error about using unknown middlewareData prop from floating-ui